workin on admin attendance

+++ aep/sections/models.py
@@ -556,9 +556,7 @@ class Attendance(models.Model):
         blank=True,
         null=True
     )
-    att_hours = models.DecimalField(
-        max_digits=4,
-        decimal_places=2,
+    att_hours = models.FloatField(
         blank=True,
         null=True
     )
